refactor: extract DataDirectoryChecker to src/refactors/data_directory_checker.py (SC-005)
Implements T049-T059. Moves the DataDirectoryChecker class (early data-directory
writable-permission check) out of MistHelper.py into a dedicated module under
src/refactors/. Unlike PR-03/PR-04, this extraction uses NO late-import DI:
the class is stdlib-only (os, sys, logging) and never touches MistHelper
globals, so the import graph stays flat.

- New module: src/refactors/data_directory_checker.py (128 LoC, 100.0/A+)
- MistHelper.py: -100 LoC net (class body removed, import added at line 149)
- Repo compliance baseline: 253 files @ 99.5/A+ preserved
- No wrapper shim; MistHelper.DataDirectoryChecker resolves via re-export

Refs: specs/1010-misthelper-refactor-extraction (PR-05)
