+++ MistHelper.py
@@ -146,6 +146,7 @@ from src.gateway.gateway_export_utils import (
     configure_gateway_export_utils_dependencies,
 )  # Import gateway export utility configuration
 from src.org_data_collector import OrgDataCollector  # Import org-level data collection orchestrator
+from src.refactors.data_directory_checker import DataDirectoryChecker  # Early data-dir writable check (SC-005)
 from src.refactors.sqlite_database_writer import SQLiteDatabaseWriter  # Extracted SQLite writer (SC-003)
 from src.refactors.tui_launcher import TUILauncher  # Extracted TUI launcher (SC-004)
 from src.reports.e911_bssid import E911BSSIDReportGenerator  # Module-level for tests
@@ -193,106 +194,6 @@ _early_log_path = os.path.join(
 # DATA DIRECTORY PERMISSION CHECKER
 # ============================================================================
 # Critical for container deployments - runs as non-root 'misthelper' user
-
-
-class DataDirectoryChecker:
-    """
-    Check data directory write permissions and provide actionable guidance.
-
-    This class runs very early during module initialization (before logging),
-    so it uses print() for output rather than logging.
-
-    Usage:
-        DataDirectoryChecker(_early_log_dir).check()
-    """
-
-    def __init__(self, data_dir: str):  # Initialize checker with target data directory path
-        """Initialize with the data directory path to check."""
-        self.data_dir = data_dir  # Store the data directory path for later validation
-        self.test_file = os.path.join(
-            data_dir, ".write_test"
-        )  # Define test file path (.write_test) for permission validation
-
-    def check(self) -> bool:  # Check if data directory is writable and handle errors
-        """Check if data directory is writable.
-
-        Returns:
-            True if writable, exits program if not writable due to permissions.
-        """
-        try:  # Attempt to validate write permission
-            return self._test_write_permission()  # Call permission test helper method
-        except PermissionError:  # Catch permission errors and display actionable guidance
-            self._handle_permission_error()  # Call error handler to print guidance and exit
-            return False  # Never reached - _handle_permission_error exits
-        except Exception:  # For non-permission errors, proceed and let them fail naturally later
-            return True  # Non-permission error, let it proceed and fail naturally
-
-    def _test_write_permission(self) -> bool:  # Validate data directory write access via test file
-        """Create and remove a test file to verify write access."""
-        with open(
-            self.test_file, "w"
-        ) as file_handle:  # Open test file for writing (will fail if directory not writable)
-            file_handle.write("test")  # Write marker content to test file
-        os.remove(self.test_file)  # Delete test file to clean up
-        return True  # Return success if both write and delete succeeded
-
-    def _handle_permission_error(self) -> None:  # Display context-specific guidance and exit
-        """Print error message with context-specific guidance and exit."""
-        in_container = self._is_running_in_container()  # Detect if running in container to show appropriate fix
-
-        self._print_error_header()  # Print error banner with path information
-
-        if in_container:  # If running in container, show container-specific fix (chmod on host)
-            self._print_container_guidance()  # Print container deployment remediation steps
-        else:  # If running locally, show local fix (chmod/chown)
-            self._print_local_guidance()  # Print local environment remediation steps
-
-        self._print_error_footer()  # Print closing separator
-        sys.exit(1)  # Exit program with error code to prevent further execution
-
-    def _is_running_in_container(self) -> bool:  # Detect container environment
-        """Detect if running inside a container environment."""
-        return os.path.exists("/.dockerenv") or os.path.exists(
-            "/run/.containerenv"
-        )  # Check for standard container marker files
-
-    def _print_error_header(self) -> None:  # Display error banner with path
-        """Print the error header with path information."""
-        print("\n" + "=" * 70)  # Print separator to visually isolate error message
-        print("ERROR: Data directory is not writable!")  # Print main error message
-        print("=" * 70)  # Print closing separator
-        print(f"\nPath: {os.path.abspath(self.data_dir)}")  # Print absolute path to the inaccessible directory
-        print("\nMistHelper cannot write logs or data to the data/ directory.")  # Explain impact of the error
-
-    def _print_container_guidance(self) -> None:  # Display container-specific remediation
-        """Print guidance specific to container deployments."""
-        print("\n[CONTAINER DETECTED]")  # Indicate container environment detected
-        print(
-            "The container runs as non-root user 'misthelper' for security."
-        )  # Explain why permissions are restricted
-        print("The mounted data/ directory must have write permissions.")  # State the requirement
-        print("\nTo fix this, run the following on your HOST machine:")  # Provide context for the fix
-        print("\n    chmod -R 777 data/")  # Show command to grant write permissions
-        print("\nThen restart the container:")  # Explain next step
-        print("    podman stop misthelper && podman rm misthelper")  # Show stop and remove command
-        print(
-            "    podman run -d --name misthelper -p 2200:2200 -p 8050:8050 \\"
-        )  # Show container restart with port mapping
-        print(
-            '        -v "${PWD}/data:/app/data:rw" -v "${PWD}/.env:/app/.env:ro" \\'
-        )  # Show volume mount with correct permissions
-        print("        ghcr.io/jmorrison-juniper/misthelper:latest")  # Show container image URI
-
-    def _print_local_guidance(self) -> None:  # Display local environment remediation
-        """Print guidance for local (non-container) environments."""
-        print("\nTo fix this, ensure the data/ directory is writable:")  # Provide context for local fix
-        print("\n    chmod -R 755 data/")  # Show command to set directory permissions
-        print("    # Or if you own the directory:")  # Provide alternative if ownership is an issue
-        print("    chown -R $(whoami) data/")  # Show command to change ownership to current user
-
-    def _print_error_footer(self) -> None:  # Display error footer separator
-        """Print the closing separator line."""
-        print("\n" + "=" * 70)  # Print separator line to visually isolate error message
 
 
 # Run data directory check immediately (NO WRAPPER - direct class instantiation)
